Extract core project models
