+++ docs/CODE_QUALITY_AUDIT.md
@@ -30,9 +30,10 @@ The architecture is moving in the right direction: core state is value typed, pe
 | `Sources/quill-code-desktop/QuillCodeDesktopApp.swift` | A- | App scene composition is now small and declarative. Keep it limited to window/menu-bar wiring and root-view routing. |
 | `Sources/quill-code-desktop/QuillCodeDesktopController.swift` | A- | Desktop controller is now mostly UI/workspace routing. Pasteboard feedback and project-import resolution now live in focused coordinators; keep future desktop protocol/workflow details out of the controller. |
 | `Sources/QuillCodeAgent/Agent.swift` | A- | Good test coverage; keep tool continuation limits and transcript filtering explicit. |
-| `Sources/QuillCodeCore/Models.swift` | A | General chat/thread/project domain models only; app config, automation scheduling, tool payloads, and TrustedRouter defaults/catalog records now live in focused core files. Watch for persistence, workflow, tool, or provider-specific behavior trying to drift back in. |
+| `Sources/QuillCodeCore/Models.swift` | A | General chat/thread/memory domain models only; app config, automation scheduling, project/workspace records, tool payloads, and TrustedRouter defaults/catalog records now live in focused core files. Watch for persistence, workflow, tool, or provider-specific behavior trying to drift back in. |
 | `Sources/QuillCodeCore/AppConfig.swift` | A | App settings, auth mode compatibility, signed-in account metadata, and favorite model normalization live together without pulling UI/runtime dependencies into core. |
 | `Sources/QuillCodeCore/AutomationModels.swift` | A | Automation kind/status/schedule records, recurrence semantics, next-run calculation, and display sorting live together without pulling app-layer automation execution into core. |
+| `Sources/QuillCodeCore/ProjectModels.swift` | A | Local/SSH project connection parsing, project refs, instructions, local environment actions, and extension manifests live together without pulling chat/thread runtime state into the project model boundary. |
 | `Sources/QuillCodeCore/ToolModels.swift` | A | Tool schema records, tool-call redaction, built-in core tool definitions, tool results, and browser/memory tool-output compatibility live together without pulling router/runtime dependencies into core. |
 | `Sources/QuillCodeCore/TrustedRouterDefaults.swift` | A | Central source of truth for TrustedRouter IDs, aliases, branded model names, fallback catalog rows, and catalog normalization. |
 | `Sources/QuillCodeCore/ModelInfo.swift` | A | Small catalog value records and sort-key semantics with no app/runtime dependency. |
@@ -97,6 +98,19 @@ Code quality changes:
 - Kept recurrence interval clamping, seconds conversion, schedule descriptions, next-run calculation, and display sorting beside the automation records.
 - Reduced `Models.swift` again so general chat/thread/project records are easier to audit independently.
 - Added a parity gate that prevents automation scheduling and display-sort rules from drifting back into `Models.swift`.
+
+## 2026-06-23 Core Project Model Ownership Pass
+
+Overall grade after this slice: **A core cohesion, A project/workspace ownership**.
+
+`Models.swift` still owned project connection parsing, SSH display, project refs, instructions, local environment actions, and extension manifests. Those are core workspace records, but they form a project boundary that should be inspectable without reading unrelated chat, approval, memory, or thread records.
+
+Code quality changes:
+
+- Moved `ProjectConnectionKind`, `ProjectConnection`, `ProjectRef`, `ProjectInstruction`, `LocalEnvironmentAction`, `ProjectExtensionKind`, `ProjectExtensionTransport`, and `ProjectExtensionManifest` into `ProjectModels.swift`.
+- Kept SSH URL/scp-style parsing, display labels, project compatibility decoding, and extension launch metadata beside the project value records.
+- Reduced `Models.swift` again so thread/message/memory records are easier to audit independently.
+- Added a parity gate that prevents project connection, SSH parsing, local action, and extension manifest records from drifting back into `Models.swift`.
 
 ## Current Refactor Priority
