+++ docs/DECISIONS.md
@@ -11,6 +11,7 @@
 - App configuration is a focused core boundary. `AppConfig.swift` owns persisted model/mode/base-URL settings, OAuth/developer-override compatibility, signed-in TrustedRouter account metadata, and favorite-model normalization; `Models.swift` stays focused on general thread/project records.
 - Core tool records are a focused schema boundary. `ToolModels.swift` owns `ToolDefinition`, `ToolCall`, `ToolResult`, built-in core tool definitions, tool-call transcript redaction, and tool-specific browser/memory output compatibility; general chat/thread/project models should not own tool payload compatibility rules.
 - Core automation records are a focused scheduling boundary. `AutomationModels.swift` owns automation kind/status/schedule enums, recurrence interval semantics, next-run calculation, and display sorting; general chat/thread/project models should not own recurring-workflow scheduling rules.
+- Core project records are a focused workspace boundary. `ProjectModels.swift` owns local/SSH connection parsing and display, project refs, instructions, local environment actions, and project extension manifests; general thread/message models should not own workspace connection or extension compatibility.
 - Tool-router feature families should delegate before they sprawl. `ToolRouter` owns the common shell/file/patch entry point and composes git definitions from `GitToolCallDispatcher`; local git, GitHub PR, and worktree tool-call argument mapping lives in that dispatcher. This keeps the shared router small while preserving one public `ToolRouter.execute` API for the agent runtime.
 - Shell request policy is a tool-family concern. `ShellToolCallDispatcher` owns `host.shell.run` definitions, cwd containment, timeout bounds, environment override validation, and `ShellExecutionRequest` construction; `ToolRouter` only delegates shell calls before handling file and patch primitives.
